backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
pymongo>=4.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
certifi>=2024.0.0
pyjwt>=2.8.0
bcrypt>=4.1.0
email-validator>=2.1.0
python-multipart>=0.0.6
